requirements.txt
pandas
streamlit
pickle
numpy
plotly.express